// src/types/allegro.ts
export interface AllegroReturn {
  id: string;
  buyer: {
    id: string;
    login: string;
  };
  parcel: {
    waybill: string;
    carrierName: string;
    carrierTrackingUrl: string;
  };
  items: {
    lineItemId: string;
    name: string;
    imageUrl?: string;
    quantity: {
      returned: number;
      unit: string;
    };
    reason: {
      id: string;
      returnReason: string;
      userComment?: string;
    };
    marketplace: {
      id: string;
    };
  }[];
  refund: {
    value: {
      amount: string;
      currency: 'PLN';
    };
    items: {
      lineItemId: string;
      lineItemValue: {
        amount: string;
        currency: 'PLN';
      };
    }[];
  };
  createdAt: string;
  updatedAt: string;
}
// ...
export interface AllegroRefund {
  id: string;
  payment: {
    id: string;
  };
  reason: {
    id: string;
    name: string;
  };
  status: 'PENDING' | 'SUCCEEDED' | 'FAILED';
  value: {
    amount: string;
    currency: 'PLN';
  };
  lineItems: {
    lineItemId: string;
    quantity: number;
    amount: {
      amount: string;
      currency: 'PLN';
    };
  }[];
  createdAt: string;
  updatedAt: string;
}